Bookmarks: Support using special charectors closes #9

+++ app/src/main/java/in/sunilpaulmathew/ashell/utils/Utils.java
@@ -15,10 +15,14 @@ import androidx.core.content.ContextCompat;
 
 import com.google.android.material.snackbar.Snackbar;
 
+import java.io.BufferedReader;
 import java.io.File;
+import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 
@@ -30,11 +34,43 @@ import in.sunilpaulmathew.ashell.R;
 public class Utils {
 
     public static boolean isBookmarked(String command, Context context) {
-        return new File(context.getExternalFilesDir("bookmarks"), command).exists();
+        if (isValidFilename(command)) {
+            return new File(context.getExternalFilesDir("bookmarks"), command).exists();
+        } else {
+            if (new File(context.getExternalFilesDir("bookmarks"), "specialCommands").exists()) {
+                for (String commands : Objects.requireNonNull(read(new File(context.getExternalFilesDir("bookmarks"), "specialCommands"))).split("\\r?\\n")) {
+                    if (commands.trim().equals(command)) {
+                        return true;
+                    }
+                }
+            }
+        }
+        return false;
     }
 
     public static boolean deleteFromBookmark(String command, Context context) {
-        return new File(context.getExternalFilesDir("bookmarks"), command).delete();
+        if (isValidFilename(command)) {
+            return new File(context.getExternalFilesDir("bookmarks"), command).delete();
+        } else {
+            StringBuilder sb = new StringBuilder();
+            for (String commands : Objects.requireNonNull(read(new File(context.getExternalFilesDir("bookmarks"), "specialCommands"))).split("\\r?\\n")) {
+                if (!commands.equals(command)) {
+                    sb.append(commands).append("\n");
+                }
+            }
+            create(sb.toString(), new File(context.getExternalFilesDir("bookmarks"), "specialCommands"));
+            return true;
+        }
+    }
+
+    /*
+     * Adapted from android.os.FileUtils
+     * Ref: https://cs.android.com/android/platform/superproject/+/master:frameworks/base/core/java/android/os/FileUtils.java;l=972?q=isValidFatFilenameChar
+     */
+    private static boolean isValidFilename(String s) {
+        return !s.contains("*") && !s.contains("/") && !s.contains(":")
+                && !s.contains("<") && !s.contains(">") && !s.contains("?")
+                && !s.contains("\\") && !s.contains("|");
     }
 
     public static Drawable getDrawable(int drawable, Context context) {
@@ -48,7 +84,16 @@ public class Utils {
     public static List<String> getBookmarks(Context context) {
         List<String> mBookmarks = new ArrayList<>();
         for (File file : Objects.requireNonNull(context.getExternalFilesDir("bookmarks").listFiles())) {
-            mBookmarks.add(file.getName());
+            if (!file.getName().equalsIgnoreCase("specialCommands")) {
+                mBookmarks.add(file.getName());
+            }
+        }
+        if (new File(context.getExternalFilesDir("bookmarks"), "specialCommands").exists()) {
+            for (String commands : Objects.requireNonNull(read(new File(context.getExternalFilesDir("bookmarks"), "specialCommands"))).split("\\r?\\n")) {
+                if (!commands.trim().isEmpty()) {
+                    mBookmarks.add(commands.trim());
+                }
+            }
         }
         return mBookmarks;
     }
@@ -63,8 +108,44 @@ public class Utils {
         return Build.MODEL;
     }
 
+    private static String read(File file) {
+        BufferedReader buf = null;
+        try {
+            buf = new BufferedReader(new FileReader(file));
+
+            StringBuilder stringBuilder = new StringBuilder();
+            String line;
+            while ((line = buf.readLine()) != null) {
+                stringBuilder.append(line).append("\n");
+            }
+
+            return stringBuilder.toString().trim();
+        } catch (IOException ignored) {
+        } finally {
+            try {
+                if (buf != null) buf.close();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
+        return null;
+    }
+
     public static void addToBookmark(String command, Context context) {
-        create(command, new File(context.getExternalFilesDir("bookmarks"), command));
+        if (isValidFilename(command)) {
+            create(command, new File(context.getExternalFilesDir("bookmarks"), command));
+        } else {
+            StringBuilder sb = new StringBuilder();
+            if (new File(context.getExternalFilesDir("bookmarks"), "specialCommands").exists()) {
+                for (String commands : Objects.requireNonNull(read(new File(context.getExternalFilesDir("bookmarks"), "specialCommands"))).split("\\r?\\n")) {
+                    sb.append(commands).append("\n");
+                }
+                sb.append(command).append("\n");
+            } else {
+                sb.append(command).append("\n");
+            }
+            create(sb.toString(), new File(context.getExternalFilesDir("bookmarks"), "specialCommands"));
+        }
     }
 
     public static void copyToClipboard(String text, Context context) {
